pyproject.toml
[project]
name = "251021"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "inference>=0.58.2",
    "inference-cli>=0.58.2",
    "inference-sdk>=0.58.2",
    "supervision>=0.26.1",
    "ultralytics>=8.3.218",
    "paho-mqtt>=1.6.1,<1.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "python-json-logger>=4.0.0",
]
